fix tests and add a previously dropped method
